+++ tests/TestHelper/TestHelper.psm1
@@ -278,12 +278,12 @@ function Get-InvalidArgumentRecord
     (
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
-        [String]
+        [System.String]
         $Message,
 
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
-        [String]
+        [System.String]
         $ArgumentName
     )
 
@@ -312,7 +312,7 @@ function Get-InvalidOperationRecord
     param
     (
         [ValidateNotNullOrEmpty()]
-        [String]
+        [System.String]
         $Message,
 
         [ValidateNotNull()]